+++ tests/testthat/test-dplyr-verbs.R
@@ -39,3 +39,19 @@ test_that("mutate and filter sample data", {
   expect_identical(nsamples(ps1), 7L)
   expect_identical(sample_data(ps1), sam1)
 })
+
+test_that("renaming taxonomy table and sample data", {
+  ps1 <- GlobalPatterns %>%
+    rename_tax_table(Domain = Kingdom) %>%
+    rename_with_tax_table(stringr::str_to_lower) %>%
+    rename_with_sample_data(janitor::make_clean_names) %>%
+    rename_sample_data(sample_id = x_sample_id)
+
+  nms <- rank_names(GlobalPatterns) %>% stringr::str_to_lower()
+  nms[1] <- "domain"
+  expect_identical(ps1 %>% rank_names, nms)
+
+  nms <- sample_variables(GlobalPatterns) %>% janitor::make_clean_names()
+  nms[1] <- "sample_id"
+  expect_identical(ps1 %>% sample_variables, nms)
+})